adc: starting 1.0.6
Added KEYP extension; note that the URI is slightly changed from the
'original' KEYP wiki entry; some wording changed to be more simple to grasp.

Note that ADCS is not added at this point in time, just yet. ADCS is not
needed at this point particularly, since the KEYP spec only relies on a)
the URI (which is simply the same as the standard URI with an 's' appended)
and b) the general use of "certificates".