adc: made editorial updates to the document
Made editorial updates to the document. However, this should create a
new version; 1.0.2. AS OF YET UNRELEASED!

Check the diff to see the differences (the only "real" difference is the
IPv6 RFC reference which is likely people have updated to anyway).

It is possible this document contain a mismatch of CRLF/LF.